Refactor File.get() to move most code to SecureSite.get_file()
